Documented a feature we support (that Rack::Builder doesn't).
The Rack::Builder and RackStack constructors both take blocks that can
be instance eval'd and can have block arguments passed to them.  Made a
spec to add coverage for us checking the arity of our block and calling
the block when an argument is passed (allowing users to not worry about
scoping issues inside of the block).